Suggest grid tool - take 2
For each track:
- Choose a default track room that has enough capacity and enough available
slots to fit all session tracks, if possible, starting with rooms that have a
maximum number of available slots. Relax capacity and slot number constraints
if there is no ideal candidate room. In practice, unless we're running short
on rooms, this should select a room for the track that is still unused.
- For each session in the track that hasn't been processed yet (a session may
belong to two tracks and thus may already have been assigned a room and slot
in a previous loop):
  - Attempt to assign a room and slot that meets all constraints. See below.
  - If that fails, relax constraints one by one and start over:
    1) stop worrying about finding a slot with the exact requested duration,
    2) stop worrying about scheduling the session in the default track room,
    3) stop worrying about even complying with the requested duration,
    4) stop worrying about the room capacity.
    5) stop worrying about session/track conflicts (partially or fully)

- Choose a default track room that has enough capacity and enough available
slots to fit all session tracks, if possible, starting with rooms that have a
maximum number of available slots. Relax capacity and slot number constraints
if there is no ideal candidate room. In practice, unless we're running short
on rooms, this should select a room for the track that is still unused.
- For each session in the track that hasn't been processed yet (a session may
belong to two tracks and thus may already have been assigned a room and slot
in a previous loop):
  - Attempt to assign a room and slot that meets all constraints. See below.
  - If that fails, relax constraints one by one and start over:
    1) stop worrying about finding a slot with the exact requested duration,
    2) stop worrying about scheduling the session in the default track room,
    3) stop worrying about even complying with the requested duration,
    4) stop worrying about the room capacity.
    5) stop worrying about session/track conflicts (partially or fully)

Now, the function that attempts to assign a room and slot for a session given
a set of constraints runs as follows:
- List possible rooms. If we explicitly set a room already, that's the only
possibility. Otherwise, if the default track room constraint is set, that's
the only possible choice. Otherwise, all rooms that have enough capacity are
possible, or all rooms if capacity constraint has been relaxed already.
- For each possible room:
  - List possible slots: If we explicitly set a slot already, that's the only
  possibility. Otherwise, all the slots that are still available in the room
  are possible. If we're dealing with a real track, we'll consider them in
  order. If we're dealing with a session that is not in a track, possible slots
  are ordered so that less used ones get considered first (to avoid gaps).
  - Find first non conflicting slot in the list, in other words a slot that
  does not lead to a situation where:
    1) Two sessions in the same track are scheduled at the same time
    2) Two sessions chaired by the same person are scheduled at the same time
    3) Conflicting sessions are scheduled at the same time
    4) Session is scheduled in a slot that does not meet its duration
    requirement
    (unless constraints have been relaxed)
  - If a suitable slot was found in the current room, we're done, otherwise
  move on to next room, or surrender for this set of constraints.

The algorithm is the same for sessions that do not belong to any track except
that the algorithm does not choose a track room to start with.