Suggest-grid.mjs generates HTML grid (#68)
* suggest-grid.mjs now generates as output an HTML file including a table,
and a list of any unscheduled sessions.

All other messages are printed via console.warn.

Highlights the following errors:
 * Two sessions from the same track scheduled in different rooms during the
   same slot.
 * When a session is scheduled in a slot with other sessions that it was
   not to conflict with.
 * When a session's declared capacity exceeds that of the room.

Note: The scheduling algorithm does not allow two sessions to be scheduled
in the same slot when they have the same chair(s). Therefore, this
code no longer checks for that.