Drop "warning: agenda" check
Part of https://github.com/w3c/tpac-breakouts/issues/53

The agenda link proved confusing and perhaps not worth including as many people
have just used the session description for the agenda or tried to push a full
agenda into what was supposed to be just a link.

This update stops issuing warnings for when the agenda is missing.