Allow additional links in materials section
The materials section is restricted to linking to the agenda, slides, minutes,
and calendar, but people are also using it to store additional links, including
links to other representations of the slides and minutes. That sounds like a
reasonable usage.

This update relaxes the check to allow all sorts of links in this section.