Ignore validation warnings if requested in Note field
This adds a new mechanism to tell the validation logic to ignore one or more
warnings when it validates a session if the project's "Note" field for the
session contains "-warning:[type]".

For instance, the validation logic will not add a "minutes" warning if the
"Note" field for the session contains "-warning:minutes".

This mechanism allows to get rid of warnings that we checked and prefer to
ignore, such as:
- capacity warnings when we chose not to follow the capacity preference
- track warnings that cannot be avoided because there are too many sessions in
the track
- minutes warnings when we know that there were no minutes taken.

The "-warning:minutes" note may also be written "-warn:minutes" or "-w:minutes".

The "Note" field may contain additional text as needed. For example:
"Slot 5 requested. No minutes as session did not take place (-warning:minutes)."